Hasta aqui solo fue claude, voy hacer pruebas con gemini

+++ Backend-AgroCredito-main/Backend-AgroCredito/src/main/java/com/AgroCredito/Model/Credito.java
@@ -40,7 +40,7 @@ public class Credito {
     private Integer plazoMeses;
     
     @Field("cuota_mensual")
-    private Double cuotaMensual;
+    private Long cuotaMensual;
     
     @Field("fecha_aprobacion")
     private LocalDateTime fechaAprobacion;
